Move openuv base entity to separate module
