+++ homeassistant/components/openuv/__init__.py
@@ -19,9 +19,6 @@ from homeassistant.const import (
 )
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import aiohttp_client
-from homeassistant.helpers.device_registry import DeviceEntryType, DeviceInfo
-from homeassistant.helpers.entity import EntityDescription
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 from .const import (
     CONF_FROM_WINDOW,
@@ -110,26 +107,3 @@ async def async_migrate_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
         LOGGER.debug("Migration to version %s successful", version)
 
     return True
-
-
-class OpenUvEntity(CoordinatorEntity):
-    """Define a generic OpenUV entity."""
-
-    _attr_has_entity_name = True
-
-    def __init__(
-        self, coordinator: OpenUvCoordinator, description: EntityDescription
-    ) -> None:
-        """Initialize."""
-        super().__init__(coordinator)
-
-        self._attr_extra_state_attributes = {}
-        self._attr_unique_id = (
-            f"{coordinator.latitude}_{coordinator.longitude}_{description.key}"
-        )
-        self.entity_description = description
-        self._attr_device_info = DeviceInfo(
-            identifiers={(DOMAIN, f"{coordinator.latitude}_{coordinator.longitude}")},
-            name="OpenUV",
-            entry_type=DeviceEntryType.SERVICE,
-        )
