+++ homeassistant/components/openuv/sensor.py
@@ -18,7 +18,6 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.util.dt import as_local, parse_datetime
 
-from . import OpenUvEntity
 from .const import (
     DATA_UV,
     DOMAIN,
@@ -34,6 +33,7 @@ from .const import (
     TYPE_SAFE_EXPOSURE_TIME_6,
 )
 from .coordinator import OpenUvCoordinator
+from .entity import OpenUvEntity
 
 ATTR_MAX_UV_TIME = "time"
